backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-decouple
requests
pyjwt
cryptography
openai
flake8 